backend/compress-images/requirements.txt
Pillow>=10.0.0
boto3
